requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.4
django-3-jet==1.0.8
django-phonenumber-field==5.0.0
idna==2.10
oauthlib==3.1.0
phonenumbers==8.12.14
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2
